Cleanup collections components
